Send certificate name derived from RENEWED_LINEAGE
Certificates issued using some of the Let's Encrypt Profiles https://letsencrypt.org/docs/profiles/ do not contain CN (Common Name). For validation, the SAN (Subject Alternative Names) extension field is used instead, but for management purposes I need a single "main" domain name. The domains in the SAN field are alphabetically sorted, and the "main" domain is not necessarily the first one. So we'll need to receive the "main" domain as well. And coincidentally, that's the directory name. Unless it's changed in the configuration to something wild, of course, which won't happen, hopefully.